legacy/requirements.txt
chromadb==1.3.6
langchain==1.1.3
langchain-chroma==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.3
langchain-google-genai==4.0.0
langchain-huggingface==1.1.0
langchain-ollama==1.0.1
langchain-text-splitters==1.0.0
pypdf==6.4.1
sentence-transformers==5.1.2
streamlit==1.52.1
python-dotenv==1.2.1